feat. 지도 일부분 (3/9)

+++ src/main/java/com/springboot/biz/root/rootUser/RootAuthList.java
@@ -16,7 +16,7 @@ public class RootAuthList {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer rootAuthListSeq;
+    private Long rootAuthListSeq;
 
     private String rootAuthListImagePath;
